[gaim-migrate @ 11144] cleaning out the 64-bit stuff from my tree...piece by piece
--HG--
extra : convert_revision : 955fea57f89edbf206e90b8810b8cf93949a117e

+++ src/protocols/oscar/oscar.c
@@ -6482,7 +6482,7 @@ static int oscar_send_chat(GaimConnection *gc, int id, const char *message) {
 	char *buf, *buf2;
 	fu16_t charset, charsubset;
 	char *charsetstr = NULL;
-	gsize len;
+	int len;
 
 	if (!(conv = gaim_find_chat(gc, id)))
 		return -EINVAL;
